flips left/right axis
left and right had the wrong polarity in the software AND the button was
connected backwards. result: motion was correct with respect to the
button-presses, position count was correct, but the device claimed to be
moving left when it was in fact moving right; this has the potential to
confuse people later.